# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChittyOS GPT Connector API",
    "description": "Comprehensive API for connecting custom GPTs to ChittyOS services, third-party integrations, and banking connections",
    "version": "1.0.0",
    "contact": {
      "name": "ChittyOS Support",
      "url": "https://chitty.cc"
    }
  },
  "servers": [
    {
      "url": "https://connect.chitty.cc",
      "description": "Production ChittyConnect API"
    },
    {
      "url": "https://connect-staging.chitty.cc",
      "description": "Staging environment"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/api/chittyid/mint": {
      "post": {
        "summary": "Mint ChittyID",
        "description": "Create a new ChittyID for an entity",
        "operationId": "mintChittyID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entity": {
                    "type": "string",
                    "enum": ["PEO", "PLACE", "PROP", "EVNT", "AUTH", "INFO", "FACT", "CONTEXT", "ACTOR"],
                    "description": "Entity type for ChittyID"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional metadata for the entity"
                  }
                },
                "required": ["entity"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ChittyID minted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChittyIDResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chittycases/create": {
      "post": {
        "summary": "Create Legal Case",
        "description": "Create a new legal case in ChittyCases",
        "operationId": "createCase",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "caseType": {
                    "type": "string",
                    "enum": ["eviction", "litigation", "resolution", "general"]
                  },
                  "metadata": {
                    "type": "object"
                  }
                },
                "required": ["title", "caseType"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Case created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caseId": {
                      "type": "string"
                    },
                    "chittyid": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chittycontextual/analyze": {
      "post": {
        "summary": "Analyze Context",
        "description": "Perform contextual analysis on provided data",
        "operationId": "analyzeContext",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "context": {
                    "type": "object"
                  },
                  "analysisType": {
                    "type": "string",
                    "enum": ["sentiment", "entities", "legal", "financial", "comprehensive"]
                  }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "object"
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "entities": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chittyevidence/ingest": {
      "post": {
        "summary": "Ingest Evidence",
        "description": "Ingest and process evidence files",
        "operationId": "ingestEvidence",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "caseId": {
                    "type": "string"
                  },
                  "evidenceType": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evidence ingested",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "evidenceId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "extractedData": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/thirdparty/notion/query": {
      "post": {
        "summary": "Query Notion Database",
        "description": "Query Notion database via ChittyConnect proxy",
        "operationId": "queryNotion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "databaseId": {
                    "type": "string"
                  },
                  "filter": {
                    "type": "object"
                  },
                  "sorts": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                },
                "required": ["databaseId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ChittyOS-API-Key"
      }
    },
    "schemas": {
...
      "ChittyIDResponse": {
        "type": "object",
        "properties": {
          "chittyid": {
            "type": "string",
            "description": "Generated ChittyID"
          },
          "entity": {
            "type": "string"
          },
          "sequence": {
            "type": "string"
          },
          "checksum": {
            "type": "string"
          }
        }
      },
...
    }
  }
}